agent-scheduler/requirements.txt
# Agent Scheduler Brain Dependencies

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration parsing
pyyaml>=6.0

# Database
psycopg2-binary>=2.9.0

# Data validation
pydantic>=2.0.0

# qwen-agent framework
qwen-agent>=0.0.3

# HTTP client
httpx>=0.25.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.90.0
pytest-asyncio>=0.21.0
testcontainers>=3.7.0
